Very good

+++ server/src/routes/comments.js
@@ -19,7 +19,8 @@ router.get(
                 "users.username",
                 "comments.id",
                 "comments.body",
-                "comments.created_at"
+                "comments.created_at",
+                "comments.user_id"
             )
             .join("users", "users.id", "comments.user_id")
             .where("comments.post_id", postId)
@@ -69,23 +70,41 @@ router.delete(
     "/deletecomment",
     authenticate,
     asyncHandler(async (req, res) => {
-        const { comment_id } = req.body;
+        const { comment_id, user_id } = req.body;
+        if (
+            !Number.isInteger(comment_id) ||
+            !Number.isInteger(user_id) ||
+            comment_id <= 0 ||
+            user_id <= 0
+        ) {
+            return res.status(400).json({
+                message: "Invalid comment_id or user_id",
+            });
+        }
+        if (req.userId !== user_id) {
+            return res.status(403).json({
+                message: "You can't delete comments by other users",
+            });
+        }
         try {
             const deletedComment = await Comments.query().deleteById(
                 comment_id
             );
+
             if (deletedComment) {
                 res.status(200).json({
-                    message: "Comment successfully deleted",
+                    message: "Comment deleted successfully",
                 });
             } else {
                 res.status(404).json({
-                    message: "Commnent not found",
+                    message: "Comment not found or unauthorized",
                 });
             }
         } catch (error) {
             console.error(error);
-            res.status(500).json({ message: "Failed to delete comment" });
+            res.status(500).json({
+                message: "Internal Server Error",
+            });
         }
     })
 );
